ingestion_layer/requirements.txt
kafka-python==2.0.2
requests==2.32.3
fastapi==0.111.0
prometheus-client==0.20.0